test: colocate core and knowledge tests
