+++ src/core/tests/conftest.py
@@ -1,0 +1,1 @@
+"""Pytest fixtures for the core test suite."""
